Set SELinux to permissive in the DbProcessor
As of Marshmallow it looks like SELinux is getting in the way of
accessing the files we copied from the contacts directory. Setting
SELinux to permissive gets the processor working again. It is set back
to enforcing at the end of the task, if necessary, to avoid leaving
the device in a less-secure state.